third_party/WebKit/LayoutTests/fast/events/touch/gesture/synthetic-pinch-zoom-gesture.html

<!DOCTYPE html>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<meta name="viewport" content="width=device-width">
<style>
  body, html {
    width: 100%;
    height: 100%;
    margin: 0;
  }
  div {
    background-color: green;
    width: 100px;
    height: 100px;
    position: absolute;
    left: 50%;
    top: 50%;
    margin-left: -50px;
    margin-top: -50px;
  }
</style>

<script>
  window.addEventListener('load', () => {
    test(function(t) {
      assert_not_equals(window.chrome.gpuBenchmarking, undefined, 'This test requires gpuBenchmarking.');

      window.chrome.gpuBenchmarking.pinchBy(4, 0.5, 0.5, () => {
        assert_equals(window.visualViewport.scale, 4);
      }, 1000);

    }, 'This tests that gpuBenchmarking.pinchBy is relatively accurate.');
  }
</script>
